hw_TG_bot fix done

+++ src/main/resources/liquibase/scripts/notification_task.sql
@@ -4,7 +4,7 @@
 create table notification_task
 (
     task_id   serial primary key,
-    chat_id   serial    not null,
+    chat_id bigint not null,
     text      text      not null,
     date_time timestamp not null
 );
